# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aozora RAG API",
    "version": "0.1.0"
  },
  "paths": {
    "/chat/": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Chat Endpoint",
        "operationId": "chat_endpoint_chat__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload": {
      "post": {
        "summary": "Upload File",
        "description": "ファイルをアップロードし、ドキュメントとして処理します。\n\n- **file**: アップロードするファイル\n- **title**: ユーザーが指定するタイトル",
        "operationId": "upload_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerItems": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SourceItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AnswerItems"
      },
      "Body_upload_file": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "file",
          "title"
        ],
        "title": "Body_upload_file"
      },
      "ChatInput": {
        "properties": {
          "questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions",
            "examples": [
              [
                "質問1",
                "質問2",
                "質問3"
              ]
            ]
          }
        },
        "type": "object",
        "title": "ChatInput"
      },
      "ChatOutput": {
        "properties": {
          "responses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnswerItems"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responses"
          }
        },
        "type": "object",
        "title": "ChatOutput",
        "description": "チャットボットの出力データの仕様書"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SourceItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url",
          "page"
        ],
        "title": "SourceItem"
      },
      "UploadResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "processed_pages": {
            "type": "integer",
            "title": "Processed Pages"
          }
        },
        "type": "object",
        "required": [
          "message",
          "processed_pages"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}